requirements.txt
aioredis==1.2.0
amqp==2.5.1
asgiref==3.2.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.4
djangorestframework==3.10.2
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.19
incremental==17.5.0
kombu==4.6.4
msgpack==0.6.1
pkg-resources==0.0.0
psycopg2==2.8.3
pycparser==2.19
PyHamcrest==1.9.0
pytz==2019.2
redis==3.3.8
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.3
vine==1.3.0
zipp==0.5.2
zope.interface==4.6.0
